Derive a custom Lindat version of token connect plugin
... from default_token_connect with as few redundance as possible.
Must be tested!
