+++ public/files/js/plugins/defaultTokenConnect/style.less
@@ -32,55 +32,4 @@
         text-decoration: underline;
     }
 
-    // Vallex service --->
-
-    div.VallexJsonRenderer {
-
-        ul.vallexHiddenBullets {
-            list-style-type: none;
-            margin: 0;
-            padding: 0;
-        }
-
-        a.vallexSense {
-            margin: 0 0 0.5em 0;
-            font-size: 16px;
-            font-weight: bold;
-            padding: 0;
-            cursor: pointer;
-        }
-
-        div.vallexSourceV {
-            font-weight: bold;
-        }
-
-        .vallexFrame {
-            color: darkred;
-            margin: 0 0 0 0.2em;
-        }
-
-        div.vallexExpl {
-            color: #777;
-            font-weight: bold;
-        }
-
-        li.vallexExamples {
-            margin: 0 0 0.2em 0;
-            padding: 0;
-        }
-
-        div.vallexTargetBlock {
-            margin: 0 0 1.4em 0.8em;
-
-            div.vallexTargetV {
-                font-weight: bold;
-            }
-
-            div.vallexExplInner {
-                color: #777;
-                font-weight: bold;
-            }
-        }
-
-    }
 }
