+++ public/files/js/plugins/lindatTokenConnect/vallex.ts
